@@ -1,7 +1,5 @@
-﻿using System;
+﻿
 using System.Text.RegularExpressions;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace Task5._2
 {
@@ -13,20 +11,20 @@ namespace Task5._2
 
         public void Add(string key, Book value)
         {
-            if (_catalogue.ContainsKey(key))
-            {
-                throw new ArgumentException();
-            }
-
             if (value == null || key == null)
             {
                 throw new ArgumentNullException();
             }
 
+            if (_catalogue.ContainsKey(key))
+            {
+                throw new ArgumentException();
+            }
+
             var ISBN = ISBNCheck(key);
             _catalogue.Add(ISBN, value);
 
-            foreach(var val in value.Authors)
+            foreach(var val in value.GetAuthors())
             {
                 _authors.Add(val);
             }
@@ -44,46 +42,28 @@ namespace Task5._2
             }
         }
 
-        public Book GetBook(string key)
+        public Book? GetBook(string key)
         {
             key = ISBNCheck(key);
-            if(_catalogue.ContainsKey(key))
-            {
-                return _catalogue.Where(x => x.Key == key).Select(x => x.Value).First();
 
-            }
-            else
-            {
-                throw new ArgumentOutOfRangeException();
-
-            }
+            return (_catalogue.ContainsKey(key)) ? _catalogue.Where(x => x.Key == key).Select(x => x.Value).First() : null;
         }
 
         public IEnumerable<string> AlphabitOrder()
         {
-            var books = _catalogue.OrderBy(x => x.Value.Title).Select(x => x.Value.Title);
-
-            foreach(var b in books)
-            {
-                yield return b;
-            }
+            return _catalogue.OrderBy(x => x.Value.Title).Select(x => x.Value.Title);
         }
 
         public IEnumerable<string> BooksList(string author)
         {
-            var books = _catalogue.Where(x => x.Value.Authors.Contains(author)).OrderBy(x => x.Value.PublishingDate).Select(x => x.Value);
-
-            foreach (var a in books)
-            {
-                 yield return a.ToString();
-            }
+            return _catalogue.Where(x => x.Value.GetAuthors().Contains(author)).OrderBy(x => x.Value.PublishingDate).Select(x => x.Value.ToString());
         }
 
         public IEnumerable<(string, int)> BooksAmount()
         {
             foreach (var author in _authors)
             {
-                int count = _catalogue.Values.Count(auth => auth.Authors.Contains(author.ToString()));
+                int count = _catalogue.Values.Count(auth => auth.GetAuthors().Contains(author.ToString()));
                 yield return (author.ToString(), count);
             }
         }
